ISPN-6021 Merge DoubleSum and DoubleAvg classes
